+++ plugin/pieces.lua
@@ -1,5 +1,5 @@
 require("pieces.copilot")
 require("pieces.assets")
 require("pieces.copilot.slash_commands")
-require("pieces.feedbacks")
+require("pieces.feedback")
 require("pieces.tutor")
